Create a mod for range_check and make sure rc_20 hardcodes 20 everywhere
